[refactor] service layer 를 application layer로 이동
