+++ src/main/java/junseok/snr/wallet/application/service/WalletService.java
@@ -1,4 +1,4 @@
-package junseok.snr.wallet.api.service;
+package junseok.snr.wallet.application.service;
 
 import junseok.snr.wallet.api.controller.dto.CreateWalletRequestDto;
 import junseok.snr.wallet.domain.Wallet;
